requirements.txt
gradio
deepgram-sdk
python-dotenv
elevenlabs
openai
langchain